Progress with test improvements
